TIP-706: Fix comments
