+++ src/Pim/Bundle/CatalogBundle/Elasticsearch/Filter/AbstractFieldFilter.php
@@ -2,19 +2,26 @@
 
 namespace Pim\Bundle\CatalogBundle\Elasticsearch\Filter;
 
+use Pim\Bundle\CatalogBundle\Elasticsearch\SearchQueryBuilder;
 use Pim\Component\Catalog\Query\Filter\FieldFilterInterface;
 
 /**
- * Basic implementation of a field Filter
+ * Abstract field filter for Elasticsearch
  *
  * @author    Samir Boulil <samir.boulil@akeneo.com>
  * @copyright 2017 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-abstract class AbstractFieldFilter extends AbstractFilter implements FieldFilterInterface
+abstract class AbstractFieldFilter implements FieldFilterInterface
 {
+    /** @var SearchQueryBuilder */
+    protected $searchQueryBuilder = null;
+
     /** @var array */
-    protected $supportedFields;
+    protected $supportedFields = [];
+
+    /** @var array */
+    protected $supportedOperators = [];
 
     /**
      * {@inheritdoc}
@@ -27,8 +34,38 @@ abstract class AbstractFieldFilter extends AbstractFilter implements FieldFilter
     /**
      * {@inheritdoc}
      */
+    public function supportsOperator($operator)
+    {
+        return in_array($operator, $this->supportedOperators);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getOperators()
+    {
+        return $this->supportedOperators;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
     public function getFields()
     {
         return $this->supportedFields;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setQueryBuilder($searchQueryBuilder)
+    {
+        if (!$searchQueryBuilder instanceof SearchQueryBuilder) {
+            throw new \InvalidArgumentException(
+                sprintf('Query builder should be an instance of "%s"', SearchQueryBuilder::class)
+            );
+        }
+
+        $this->searchQueryBuilder = $searchQueryBuilder;
     }
 }
