--- src/Pim/Bundle/CatalogBundle/Elasticsearch/Filter/AbstractFilter.php
@@ -1,52 +1,0 @@
-<?php
-
-namespace Pim\Bundle\CatalogBundle\Elasticsearch\Filter;
-
-use Pim\Bundle\CatalogBundle\Elasticsearch\SearchQueryBuilder;
-use Pim\Component\Catalog\Query\Filter\FilterInterface;
-
-/**
- * Basic implementation of a field filter.
- *
- * @author    Samir Boulil <samir.boulil@akeneo.com>
- * @copyright 2016 Akeneo SAS (http://www.akeneo.com)
- * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
- */
-abstract class AbstractFilter implements FilterInterface
-{
-    /** @var SearchQueryBuilder */
-    protected $searchQueryBuilder = null;
-
-    /** @var array */
-    protected $supportedOperators = [];
-
-    /**
-     * {@inheritdoc}
-     */
-    public function supportsOperator($operator)
-    {
-        return in_array($operator, $this->supportedOperators);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getOperators()
-    {
-        return $this->supportedOperators;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setQueryBuilder($searchQueryBuilder)
-    {
-        if (!$searchQueryBuilder instanceof SearchQueryBuilder) {
-            throw new \InvalidArgumentException(
-                sprintf('Query builder should be an instance of "%s"', SearchQueryBuilder::class)
-            );
-        }
-
-        $this->searchQueryBuilder = $searchQueryBuilder;
-    }
-}
